library management system
Library  Management System is an windows based application that was developed and designed with the sole objective of automating library services. 
This system aimed at automating the processes of cataloging, book circulation, Students details maintenance, and Record maintenance. The system’s performance meets user’s requirements, hence providing the main benefit of concentrating all the library services and functions within the power of a mouse click and feeding of data into the system.